ngo-fraud-detection-system/production-v1/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
rapidfuzz>=3.9.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.4.0
reportlab>=4.2.0
